SP10 Framework: GenericSerializer<T> refactoring
move lookup of registered type serializer to start of process. read type header now hydrates array type so serializer can be loaded. use ushort for object index in circular ref.

+++ src/Sphere10.Framework/Serialization/GenericItemSerializer.cs
@@ -63,7 +63,7 @@ namespace Sphere10.Framework {
 		protected struct CircularReference {
 			public int TypeIndex { get; set; }
 
-			public int Index { get; set; }
+			public ushort Index { get; set; }
 		}
